requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.65.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
groq>=0.13.0
